NOJIRA: Refactor build scripts and fix local macOS binary codesign
Move build-binary and copy-resources to TypeScript, add macOS ad-hoc codesigning to fix SIGKILL on arm64, and clean up package.json scripts (deduplicate lint/typecheck, use bun consistently as task runner).
