+++ .github/workflows/release.yml
@@ -56,24 +56,11 @@ jobs:
       - uses: actions/checkout@v4
       - uses: ./.github/actions/setup
 
-      - name: Build TypeScript
-        run: pnpm run build
-
       - name: Setup Bun
         uses: oven-sh/setup-bun@0c5077e51419868618aeaa5fe8019c62421857d6 # v2
 
-      - name: Compile standalone binary
-        run: bun build src/cli.ts --compile --outfile dist/kimchi-code --external chromium-bidi --external electron
-
-      # Bun build --compile produces binaries with an invalid code signature on macOS.
-      # The kernel kills badly-signed arm64 binaries immediately (SIGKILL, exit 137).
-      # Strip the corrupt signature and re-sign ad-hoc. See: https://github.com/oven-sh/bun/issues/7208
-      - name: Ad-hoc sign binary (macOS)
-        if: runner.os == 'macOS'
-        run: codesign --remove-signature dist/kimchi-code && codesign -s - dist/kimchi-code
-
-      - name: Copy assets
-        run: pnpm run copy:assets
+      - name: Build binary
+        run: pnpm run build:binary
 
       - name: Package binary
         run: tar -czf kimchi-code_${{ matrix.os }}_${{ matrix.arch }}.tar.gz -C dist kimchi-code package.json theme
