+++ package.json
@@ -9,13 +9,13 @@
   },
   "scripts": {
     "clean": "rm -rf dist",
-    "copy:assets": "cp package.json dist/ && cp -r src/modes/interactive/theme dist/theme",
-    "build": "tsc --noEmit && node scripts/copy-resources.js",
-    "build:binary": "pnpm run clean && pnpm run build && bun build src/cli.ts --compile --outfile dist/kimchi-code --external chromium-bidi --external electron && pnpm run copy:assets",
+    "build": "tsc --noEmit && bun run scripts/copy-resources.ts --dev",
+    "build:binary": "bun run scripts/build-binary.ts",
     "dev": "bun run src/cli.ts",
-    "lint": "biome check src/",
-    "lint:fix": "biome check --write src/",
-    "check": "biome check src/ && tsc --noEmit",
+    "lint": "biome check src/ scripts/",
+    "lint:fix": "biome check --write src/ scripts/",
+    "typecheck": "tsc --noEmit",
+    "check": "pnpm run lint && pnpm run typecheck",
     "test": "vitest run --dir src",
     "test:smoke": "vitest run --config tests/smoke/vitest.config.ts",
     "verify": "pnpm run check && pnpm run build:binary && pnpm run test && pnpm run test:smoke"
