--- scripts/copy-resources.js
@@ -1,30 +1,0 @@
-// Copy non-TypeScript resources that tsc doesn't handle.
-//
-// Theme files: pi-mono resolves themes at <packageDir>/src/modes/interactive/theme/
-// so we copy them from node_modules into our source tree.
-
-import { cpSync, mkdirSync } from "node:fs"
-import { dirname, join } from "node:path"
-import { fileURLToPath } from "node:url"
-
-const __dirname = dirname(fileURLToPath(import.meta.url))
-const projectRoot = join(__dirname, "..")
-
-// ── Theme files (node_modules → src/) ────────────────────────────────
-const themeSrc = join(
-	projectRoot,
-	"node_modules",
-	"@mariozechner",
-	"pi-coding-agent",
-	"dist",
-	"modes",
-	"interactive",
-	"theme",
-)
-const themeDest = join(projectRoot, "src", "modes", "interactive", "theme")
-
-mkdirSync(themeDest, { recursive: true })
-
-for (const file of ["dark.json", "light.json", "theme-schema.json"]) {
-	cpSync(join(themeSrc, file), join(themeDest, file))
-}
